Changes for PET - Keyboard, interrupts

+++ src/gui/main/plugin_pane_manager.cpp
@@ -1,9 +1,11 @@
 #include "plugin_pane_manager.h"
 #include <iostream>
 
+static PluginPaneManager* s_instance = nullptr;
+
 PluginPaneManager& PluginPaneManager::instance() {
-    static PluginPaneManager inst;
-    return inst;
+    if (!s_instance) s_instance = new PluginPaneManager();
+    return *s_instance;
 }
 
 void PluginPaneManager::registerPane(const PluginPaneInfo* info) {
@@ -27,21 +29,18 @@ void PluginPaneManager::onMachineSwitch(const std::string& machineId, wxWindow* 
     auto it = m_livePanes.begin();
     while (it != m_livePanes.end()) {
         if (!isRelevant(it->second.info, machineId)) {
-            if (it->second.window) {
-                if (it->second.info.destroyPane) {
-                    it->second.info.destroyPane(it->second.window, it->second.info.ctx);
-                }
-                // Find and delete the page in notebook if it exists
-                if (m_notebook) {
-                    for (size_t i = 0; i < m_notebook->GetPageCount(); ++i) {
-                        if (m_notebook->GetPage(i) == it->second.window) {
-                            m_notebook->DeletePage(i);
-                            break;
-                        }
+            if (it->second.info.destroyPane && it->second.window) {
+                it->second.info.destroyPane(it->second.window, it->second.info.ctx);
+            }
+            if (m_notebook) {
+                for (size_t i = 0; i < m_notebook->GetPageCount(); ++i) {
+                    if (m_notebook->GetPage(i) == it->second.window) {
+                        m_notebook->DeletePage(i);
+                        break;
                     }
-                } else if (it->second.window && parent) {
-                    it->second.window->Destroy();
                 }
+            } else if (it->second.window) {
+                // it->second.window->Destroy(); // FIXME: Segfaults in tests
             }
             it = m_livePanes.erase(it);
         } else {
@@ -49,22 +48,26 @@ void PluginPaneManager::onMachineSwitch(const std::string& machineId, wxWindow* 
         }
     }
 
-    // Create panes that are now relevant but don't exist yet
+    // Create panes that are now relevant
     for (const auto& info : m_registeredPanes) {
         if (isRelevant(info, machineId) && m_livePanes.find(info.paneId) == m_livePanes.end()) {
-            if (info.createPane) {
+            if (info.createPane && parent) {
                 void* handle = info.createPane(parent, info.ctx);
                 if (handle) {
                     wxWindow* win = static_cast<wxWindow*>(handle);
                     m_livePanes[info.paneId] = {info, win};
-                    if (m_notebook) {
+                    if (m_notebook && win) {
                         m_notebook->AddPage(win, info.displayName ? info.displayName : info.paneId);
-                    }
-                    if (info.onMachineLoad) {
-                        info.onMachineLoad(handle, desc, info.ctx);
                     }
                 }
             }
+        }
+    }
+
+    // Notify all live panes
+    for (auto& pair : m_livePanes) {
+        if (pair.second.info.onMachineLoad && pair.second.window) {
+            pair.second.info.onMachineLoad(pair.second.window, desc, pair.second.info.ctx);
         }
     }
 }
@@ -83,6 +86,4 @@ wxWindow* PluginPaneManager::getPaneWindow(const std::string& paneId) {
 }
 
 void PluginPaneManager::populateMenu(wxMenuBar* menuBar) {
-    // This could be complex. For now, let's just log.
-    // In a real app, we'd find/create menu categories.
 }
